Splitted Main into components.
todo: split Quizsite into components and add routes.
