+++ nat-quiz/index.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+</head>
+<style>
+    .test {
+  background-color: white;
+  border-style: solid;
+  border-width: 2px;
+  border-color: red;
+  padding: 10px;
+  margin: 20px;
+  display: inline;
+}
+</style>
+<body>
+    <p class="test">test</p>
+</body>
+</html>
