+++ nat-quiz/package.json
@@ -13,6 +13,7 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "react": "^17.0.0 || ^18.0.0",
+    "react-confetti": "^6.1.0",
     "react-dom": "^17.0.0 || ^18.0.0",
     "react-scripts": "5.0.1",
     "styled-components": "^5.3.11",
@@ -49,6 +50,7 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^5.0.0"
+    "gh-pages": "^5.0.0",
+    "react-router-dom": "^6.14.2"
   }
 }
