+++ nat-quiz/src/App.css
@@ -8,66 +8,53 @@ body {
 
 .efe{
   color: whitesmoke;
-
   display: flex;
   justify-content: center;
   margin-top: 90px;
   margin-right: 20px;
   font-size: 19px;
-  
 }
 
-.logo_nat{
-  width: auto;
-  height: 69px;
-  position: absolute;
-  margin-top: -69px;
-  
-  left: 92.98%;
-  /* bottom: 80%; */
-}
-
-.h1{
+h1{
   color: white;
   display: flex;
   justify-content: center;
   margin-bottom: -45px;
 }
 
-.App .image-container img:hover {
-  transform: scale(1.1);
-}
-
 .container {
   display: flex;
-  justify-content: center;
+  flex-direction: column;
+  gap: 32px;
+  justify-content: space-around;
   align-items: center;
+  margin-top: 100px;
 }
 
-.image-container {
-  margin: 10px;
-  text-align: center;
+.quiz-item {
+  display: flex;
+  flex-direction: column;
+  height: 300px;
 }
 
 .pic {
-  width: auto;
-  height: 19vw;
   cursor: pointer;
-  margin-top: 200px;
-  margin-bottom: 30px;
-}
+  height: 100%;
+  transition: transform 0.3s ease-in-out;
+  display: block;
 
-.lbl {
-  background-color: white;
-  padding: 10px;
-  font-weight: bold;
-  margin-top: -30px;
-  z-index: 1;
-  position: relative;
 }
 
 .pic:hover {
   transform: scale(1.01);
+}
+
+.lbl {
+  background-color: white;
+  padding: 7px;
+  font-weight: bold;
+  z-index: 1;
+  position: relative;
 }
 
 .App {
@@ -85,62 +72,48 @@ body {
   height: 100%;
 }
 
-.App .image-container {
-  display: flex;
-  justify-content: center;
-  width: 100%;
-  margin-bottom: 20px;
-}
-
-.App .image-container img {
-  width: 200px;
-  height: 200px;
-  border-radius: 50%;
-  cursor: pointer;
-  margin: 0 10px;
-  transition: transform 0.3s ease-in-out;
-}
-
 /* Media Queries */
 
-@media (max-width: 768px) {
-  .pic {
-    height: 200px;
-    margin-top: 50px;
-  }
-
-  .App .image-container {
-    flex-wrap: wrap;
-    justify-content: center;
-    align-items: center;
-  }
-
-  .App .image-container img {
-    width: 150px;
-    height: 150px;
-    margin: 10px;
-  }
-}
-
-@media (max-width: 480px) {
+@media (min-width: 480px) {
   body {
     font-size: 24px;
   }
 
   .pic {
-    height: 150px;
     margin-top: 30px;
   }
 
-  .App .image-container {
+  .image-container {
     flex-wrap: wrap;
     justify-content: center;
     align-items: center;
   }
 
-  .App .image-container img {
+  .image-container img {
     width: 100px;
     height: 100px;
+    margin: 10px;
+  }
+}
+
+@media (min-width: 768px) {
+  .container {
+    flex-direction: row;
+  }
+
+  .pic {
+    margin-top: 50px;
+  }
+
+  .image-container {
+    flex-wrap: wrap;
+    justify-content: center;
+    align-items: center;
+  }
+
+  .image-container img {
+    width: 150px;
+    height: 150px;
     margin: 10px;
   }
 }
