+++ nat-quiz/src/App.jsx
@@ -6,54 +6,14 @@ import "./App.css";
 import Allgemein from "./Allgemein";
 import Tier from "./Tier";
 import Flaggen from "./Flaggen";
-
-const root = ReactDOM.createRoot(document.getElementById("root"));
-
-function allgemeinQuiz() {
-  root.render(<Allgemein />);
-}
-
-function flaggenQuiz() {
-  root.render(<Flaggen />);
-}
-
-function ZliQuiz() {
-  root.render(<Tier />);
-}
+import { Main } from "./components/Main"
 
 function App() {
   return (
     <>
-      <h1 className="h1">Willkommen im NÄT_Quiz</h1>
-
-      <div className="container">
-        <div className="image-container">
-          <img
-            onClick={flaggenQuiz}
-            className="pic"
-            src="https://cdn.pixabay.com/photo/2013/07/12/19/18/world-154527_1280.png"
-          ></img>
-          <label className="lbl">Flaggen Quiz</label>
-        </div>
-        <div className="image-container">
-          <img
-            onClick={allgemeinQuiz}
-            className="pic"
-            src="https://cdn.pixabay.com/photo/2017/01/31/00/09/books-2022464_1280.png"
-          ></img>
-          <label className="lbl">Allgemeinwissen Quiz</label>
-        </div>
-        <div className="image-container">
-          <img
-            onClick={ZliQuiz}
-            className="pic"
-            src="https://cdn.pixabay.com/photo/2016/03/31/22/14/blue-1296931_1280.png"
-          ></img>
-          <label className="lbl">Tier Quiz</label>
-        </div>
-      </div>
-      <h5 className="efe">by Deniz</h5>
-    </>
+     <Main />
+     </>
+    
   );
 }
