+++ nat-quiz/src/components/Main.jsx
@@ -1,0 +1,14 @@
+import {QuizSection} from "./QuizSection"
+import "../App.css"
+
+export const Main = () => {
+    return ( 
+    <>
+    <h1>Willkommen im NÄT-Quiz</h1>
+    <QuizSection />
+    <h5 className="efe">by Deniz</h5>
+
+    </>
+    )
+
+}
