+++ nat-quiz/src/components/Test.jsx
@@ -1,0 +1,3 @@
+export const Test = (props) => {
+    return <p style={{color: "white"}}>Hello {props.name}</p>;
+}
