+++ nat-quiz/src/index.js
@@ -4,6 +4,19 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+
+export default function App () {
+  return (
+    <BrowserRouter>
+        <Routes>
+          <Route>
+            
+          </Route>
+        </Routes>
+    </BrowserRouter>
+  )
+}
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
